api/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyPDF2==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
google-generativeai==0.4.1
